future proofing
Changed from 32 px for a tile to 16 px. It gave the constants more flexibility using variables to make it easier to change values. I made a struct to hold screen widths and height for screen config. This allows it to be changed based on one's preference. Lastly, all raw pointers were replaced with shared pointers to reduce the risk of memory leaks and stack overflow.